direct loading generified and made testable

+++ vector-etl/src/main/scala/io/hgis/domain/osm/Way.scala
@@ -21,6 +21,6 @@ class Way extends TWay {
 
   @Id
   @Column(name = "id")
-  var id: Int = _
+  var id: Long = _
 
 }
