+++ vector-etl/src/main/scala/io/hgis/domain/osm/WayWMS.scala
@@ -39,6 +39,6 @@ class WayWMS extends TWay {
 
   @Id
   @Column(name = "id")
-  var id: Int = _
+  var id: Long = _
 
 }
